refactor: moved RasterAlgorithmExplore to map/data folder of geohub
