+++ .changeset/spicy-days-scream.md
@@ -1,0 +1,5 @@
+---
+"geohub": patch
+---
+
+refactor: moved RasterAlgorithmExplore to map/data folder of geohub
