+++ sites/geohub/src/components/pages/map/data/RasterAlgorithmExplorer.svelte
