+++ sites/geohub/src/routes/(app)/tools/+page.server.ts
@@ -1,4 +1,4 @@
-import { ALGORITHM_TAG_KEY } from '$components/maplibre/raster/RasterAlgorithmExplorer.svelte';
+import { ALGORITHM_TAG_KEY } from '$components/pages/map/data/RasterAlgorithmExplorer.svelte';
 import type { DatasetFeatureCollection, Tag } from '$lib/types';
 import type { PageServerLoad } from './$types';
 import { error } from '@sveltejs/kit';
